% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mat.R
\name{load_mat}
\alias{load_mat}
\title{Load data as a matrix
source: https://tinyheero.github.io/jekyll/update/2015/07/26/making-your-first-R-package.html}
\usage{
load_mat(infile)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
This function loads a file as a matrix. It assumes the first column contains
the rownames and the subsequent columns are the sample identifiers. Rows with
duplicated rownames will be dropped with the first one being kept.
}
